creates component for JoinPoll
